refactor pruner instances page into focused modules
