+++ apps/web/src/pages/pruner/pruner-instances-list-page.tsx
@@ -1,1414 +1,21 @@
-import {
-  MmAtGlanceCard,
-  MmAtGlanceGrid,
-  MmJobsPagination,
-  MmNeedsAttentionList,
-  MmNextStepsButton,
-  MmOverviewSection,
-  MmStatCaption,
-  MmStatTile,
-  MmStatTileRow,
-} from "../../components/overview/mm-overview-cards";
-import { useEffect, useId, useMemo, useRef, useState } from "react";
-import { Link } from "react-router-dom";
-import { useQueryClient } from "@tanstack/react-query";
-import { MmOnOffSwitch } from "../../components/ui/mm-on-off-switch";
-import {
-  MM_SCHEDULE_TIME_WINDOW_HELPER,
-  MmScheduleDayChips,
-  MmScheduleTimeFields,
-} from "../../components/ui/mm-schedule-window-controls";
+import { useState } from "react";
 import {
   mmModuleTabBlurbBandClass,
   mmModuleTabBlurbTextClass,
 } from "../../lib/ui/mm-module-tab-blurb";
-import { MmListboxPicker } from "../../components/ui/mm-listbox-picker";
-import {
-  mmActionButtonClass,
-  mmSectionTabClass,
-} from "../../lib/ui/mm-control-roles";
-import { fetchCsrfToken } from "../../lib/api/auth-api";
-import type {
-  PrunerJobsInspectionRow,
-  PrunerServerInstance,
-} from "../../lib/pruner/api";
-import {
-  patchPrunerInstance,
-  patchPrunerScope,
-  postPrunerConnectionTest,
-  postPrunerInstance,
-} from "../../lib/pruner/api";
-import { useMeQuery } from "../../lib/auth/queries";
-import {
-  usePrunerInstancesQuery,
-  usePrunerJobsInspectionQuery,
-  usePrunerOverviewStatsQuery,
-} from "../../lib/pruner/queries";
-import {
-  PrunerDryRunControls,
-  PrunerProviderRulesCard,
-  type PrunerProviderRulesCardHandle,
-} from "./pruner-provider-operator-workspace";
-import {
-  formatPrunerDateTime,
-  prunerJobKindOperatorLabel,
-} from "./pruner-ui-utils";
-type TopTab = "overview" | "emby" | "jellyfin" | "plex" | "jobs" | "schedule";
-type ProviderTab = "emby" | "jellyfin" | "plex";
-
-const PRUNER_TAB_BLURBS: Record<TopTab, string> = {
-  overview:
-    "See cross-provider status for connections, cleanup coverage, and items that need attention.",
-  emby: "Manage one Emby server: test connection, set cleanup rules, and control run windows.",
-  jellyfin:
-    "Manage one Jellyfin server: test connection, set cleanup rules, and control run windows.",
-  plex: "Manage one Plex server: test connection, set cleanup rules, and control run windows.",
-  jobs: "View queued, running, and recent Pruner jobs across providers.",
-  schedule:
-    "Edit scheduled prune windows in the provider workspace and save changes per library.",
-};
-
-function providerLabel(p: ProviderTab): string {
-  if (p === "emby") return "Emby";
-  if (p === "jellyfin") return "Jellyfin";
-  return "Plex";
-}
-
-function parseServerInstanceId(job: PrunerJobsInspectionRow): number | null {
-  if (!job.payload_json) return null;
-  try {
-    const parsed = JSON.parse(job.payload_json) as {
-      server_instance_id?: unknown;
-    };
-    const sid = parsed.server_instance_id;
-    return typeof sid === "number" && Number.isFinite(sid) ? sid : null;
-  } catch {
-    return null;
-  }
-}
-
-const PRUNER_JOB_FILTER_OPTIONS = [
-  { value: "recent", label: "Recent (all statuses, newest first)" },
-  { value: "pending", label: "Pending" },
-  { value: "running", label: "Running" },
-  { value: "failed", label: "Failed" },
-  { value: "completed", label: "Completed" },
-] as const;
-
-function activeRuleCount(
-  scope: PrunerServerInstance["scopes"][number],
-): number {
-  return [
-    scope.missing_primary_media_reported_enabled,
-    scope.never_played_stale_reported_enabled,
-    scope.watched_tv_reported_enabled,
-    scope.watched_movies_reported_enabled,
-    scope.watched_movie_low_rating_reported_enabled,
-    scope.unwatched_movie_stale_reported_enabled,
-  ].filter(Boolean).length;
-}
-
-function providerCredentialLabel(provider: ProviderTab): string {
-  return provider === "plex" ? "Token" : "API key";
-}
-
-/** Placeholder when a key is stored server-side (empty field = unchanged). */
-const API_KEY_SAVED_PLACEHOLDER = "\u2022".repeat(10);
-
-function prunerConnectionDirty(
-  hasInstance: boolean,
-  savedUrl: string,
-  urlDraft: string,
-  credentialDraft: string,
-): boolean {
-  const u = urlDraft.trim();
-  const saved = (savedUrl ?? "").trim();
-  if (hasInstance) {
-    return u !== saved || credentialDraft.trim() !== "";
-  }
-  return u !== "" && credentialDraft.trim() !== "";
-}
-
-function defaultScope(scope: "tv" | "movies") {
-  return {
-    media_scope: scope,
-    missing_primary_media_reported_enabled: true,
-    never_played_stale_reported_enabled: false,
-    never_played_min_age_days: 90,
-    watched_tv_reported_enabled: scope === "tv",
-    watched_movies_reported_enabled: scope === "movies",
-    watched_movie_low_rating_reported_enabled: false,
-    watched_movie_low_rating_max_jellyfin_emby_community_rating: 4,
-    watched_movie_low_rating_max_plex_audience_rating: 4,
-    unwatched_movie_stale_reported_enabled: false,
-    unwatched_movie_stale_min_age_days: 90,
-    preview_max_items: 500,
-    preview_include_genres: [],
-    preview_include_people: [],
-    preview_include_people_roles: [],
-    preview_year_min: null,
-    preview_year_max: null,
-    preview_include_studios: [],
-    preview_include_collections: [],
-    scheduled_preview_enabled: false,
-    scheduled_preview_interval_seconds: 3600,
-    scheduled_preview_hours_limited: false,
-    scheduled_preview_days: "",
-    scheduled_preview_start: "00:00",
-    scheduled_preview_end: "23:59",
-    last_scheduled_preview_enqueued_at: null,
-    last_preview_run_uuid: null,
-    last_preview_at: null,
-    last_preview_candidate_count: null,
-    last_preview_outcome: null,
-    last_preview_error: null,
-  };
-}
-
-function providerDisabledInstance(provider: ProviderTab): PrunerServerInstance {
-  return {
-    id: 0,
-    provider,
-    display_name: `${providerLabel(provider)} (not yet connected)`,
-    base_url: "",
-    enabled: false,
-    last_connection_test_at: null,
-    last_connection_test_ok: null,
-    last_connection_test_detail: null,
-    scopes: [defaultScope("tv"), defaultScope("movies")],
-  };
-}
-
-function prunerConnectionPlaceholderUrl(provider: ProviderTab): string {
-  if (provider === "plex") return "http://localhost:32400";
-  return "http://localhost:8096";
-}
-
-function PrunerConnectionCredentialPanel({
-  provider,
-  allInstances,
-  instanceSelection,
-}: {
-  provider: ProviderTab;
-  allInstances: PrunerServerInstance[];
-  /** When set, instance row is chosen by the parent (e.g. provider workspace) so Connection matches Cleanup. */
-  instanceSelection?: {
-    selectedId: number | null;
-    onSelectedIdChange: (id: number | null) => void;
-  };
-}) {
-  const me = useMeQuery();
-  const q = useQueryClient();
-  const canOperate = me.data?.role === "admin" || me.data?.role === "operator";
-  const providerName = providerLabel(provider);
-  const providerInstances = useMemo(
-    () => allInstances.filter((x) => x.provider === provider),
-    [allInstances, provider],
-  );
-  const [internalSelectedInstanceId, setInternalSelectedInstanceId] = useState<
-    number | null
-  >(providerInstances[0]?.id ?? null);
-  const controlled = Boolean(instanceSelection);
-  const selectedInstanceId = controlled
-    ? instanceSelection!.selectedId
-    : internalSelectedInstanceId;
-  const setSelectedInstanceId = controlled
-    ? instanceSelection!.onSelectedIdChange
-    : setInternalSelectedInstanceId;
-  const selectedInstance =
-    providerInstances.find((x) => x.id === selectedInstanceId) ??
-    providerInstances[0];
-  const hasInstance = Boolean(selectedInstance);
-  const [baseUrlDraft, setBaseUrlDraft] = useState("");
-  const [credentialDraft, setCredentialDraft] = useState("");
-  const [showCredential, setShowCredential] = useState(false);
-  const [savePending, setSavePending] = useState(false);
-  const [testPending, setTestPending] = useState(false);
-  const [err, setErr] = useState<string | null>(null);
-  const [saveJustSucceeded, setSaveJustSucceeded] = useState(false);
-  const [testJustSucceeded, setTestJustSucceeded] = useState(false);
-
-  const savedUrl = selectedInstance?.base_url ?? "";
-  const dirty = prunerConnectionDirty(
-    hasInstance,
-    savedUrl,
-    baseUrlDraft,
-    credentialDraft,
-  );
-  const panelBusy = savePending || testPending;
-  const credentialPlaceholder =
-    hasInstance && credentialDraft === ""
-      ? API_KEY_SAVED_PLACEHOLDER
-      : provider === "plex"
-        ? "Enter token"
-        : "Enter API key";
-
-  const connectionStatusMain = !selectedInstance
-    ? "Not connected yet"
-    : selectedInstance.last_connection_test_ok === true
-      ? "Connected"
-      : selectedInstance.last_connection_test_ok === false
-        ? "Last test failed"
-        : "Not tested yet";
-
-  useEffect(() => {
-    if (controlled) return;
-    setInternalSelectedInstanceId(providerInstances[0]?.id ?? null);
-  }, [controlled, providerInstances.length, provider]);
-
-  useEffect(() => {
-    setBaseUrlDraft(selectedInstance?.base_url ?? "");
-    setCredentialDraft("");
-    setShowCredential(false);
-  }, [selectedInstance?.id, selectedInstance?.base_url]);
-
-  useEffect(() => {
-    if (!saveJustSucceeded) {
-      return;
-    }
-    const t = window.setTimeout(() => setSaveJustSucceeded(false), 2400);
-    return () => clearTimeout(t);
-  }, [saveJustSucceeded]);
-
-  useEffect(() => {
-    if (!testJustSucceeded) {
-      return;
-    }
-    const t = window.setTimeout(() => setTestJustSucceeded(false), 2400);
-    return () => clearTimeout(t);
-  }, [testJustSucceeded]);
-
-  async function saveConnection() {
-    setSavePending(true);
-    setErr(null);
-    setSaveJustSucceeded(false);
-    setTestJustSucceeded(false);
-    try {
-      const trimmedUrl = baseUrlDraft.trim();
-      if (!trimmedUrl) throw new Error("Base URL is required.");
-      if (!hasInstance && !credentialDraft.trim()) {
-        throw new Error(
-          `${providerCredentialLabel(provider)} is required to create a new ${providerName} connection.`,
-        );
-      }
-      const credentialKey = provider === "plex" ? "auth_token" : "api_key";
-      const credentials = credentialDraft.trim()
-        ? { [credentialKey]: credentialDraft.trim() }
-        : undefined;
-      if (selectedInstance) {
-        await patchPrunerInstance(selectedInstance.id, {
-          base_url: trimmedUrl,
-          ...(credentials ? { credentials } : {}),
-        });
-      } else {
-        await postPrunerInstance({
-          provider,
-          display_name: providerName,
-          base_url: trimmedUrl,
-          credentials: credentials ?? {},
-        });
-      }
-      await q.invalidateQueries({ queryKey: ["pruner", "instances"] });
-      setCredentialDraft("");
-      setShowCredential(false);
-      setSaveJustSucceeded(true);
-    } catch (e) {
-      setErr((e as Error).message);
-    } finally {
-      setSavePending(false);
-    }
-  }
-
-  async function runConnectionTest() {
-    if (!selectedInstance) return;
-    setTestPending(true);
-    setErr(null);
-    setSaveJustSucceeded(false);
-    setTestJustSucceeded(false);
-    try {
-      await postPrunerConnectionTest(selectedInstance.id);
-      await q.invalidateQueries({ queryKey: ["pruner", "instances"] });
-      setTestJustSucceeded(true);
-      setShowCredential(false);
-    } catch (e) {
-      setErr((e as Error).message);
-    } finally {
-      setTestPending(false);
-    }
-  }
-
-  const saveLabel = `Save ${providerName}`;
-  const testLabel = `Test ${providerName}`;
-
-  return (
-    <section
-      className={[
-        "mm-card mm-dash-card flex h-full min-h-0 min-w-0 flex-col transition-shadow duration-200",
-        saveJustSucceeded
-          ? "ring-2 ring-[var(--mm-accent-ring)] ring-offset-2 ring-offset-[var(--mm-bg-main)] shadow-[0_0_0_1px_rgba(212,175,55,0.12)]"
-          : "",
-      ].join(" ")}
-      data-testid={`pruner-connection-panel-${provider}`}
-    >
-      <div className="mm-card-action-body flex-1 min-h-0">
-        {!controlled && providerInstances.length > 1 ? (
-          <label className="block text-sm text-[var(--mm-text2)]">
-            <span className="mb-1 block text-xs text-[var(--mm-text3)]">
-              Server
-            </span>
-            <select
-              className="mm-input mt-1 w-full"
-              value={selectedInstance?.id ?? ""}
-              onChange={(e) => setSelectedInstanceId(Number(e.target.value))}
-              disabled={panelBusy || !canOperate}
-            >
-              {providerInstances.map((inst) => (
-                <option key={inst.id} value={inst.id}>
-                  {inst.display_name}
-                </option>
-              ))}
-            </select>
-          </label>
-        ) : null}
-        <label className="block text-sm text-[var(--mm-text2)]">
-          <span className="mb-1 block text-xs text-[var(--mm-text3)]">
-            Base URL
-          </span>
-          <input
-            type="url"
-            autoComplete="off"
-            value={baseUrlDraft}
-            onChange={(e) => setBaseUrlDraft(e.target.value)}
-            disabled={panelBusy || !canOperate}
-            placeholder={prunerConnectionPlaceholderUrl(provider)}
-            className="mm-input w-full"
-          />
-        </label>
-
-        <div className="space-y-1">
-          <label
-            className="block text-sm text-[var(--mm-text2)]"
-            htmlFor={`pruner-conn-key-${provider}`}
-          >
-            <span className="mb-1 block text-xs text-[var(--mm-text3)]">
-              {providerCredentialLabel(provider)}
-            </span>
-          </label>
-          <div className="flex flex-wrap gap-2">
-            <input
-              id={`pruner-conn-key-${provider}`}
-              type={showCredential ? "text" : "password"}
-              autoComplete="new-password"
-              placeholder={credentialPlaceholder}
-              className="mm-input min-w-[12rem] flex-1 text-sm tracking-normal text-[var(--mm-text)]"
-              disabled={panelBusy || !canOperate}
-              value={credentialDraft}
-              aria-describedby={
-                hasInstance ? `pruner-conn-key-hint-${provider}` : undefined
-              }
-              onChange={(e) => {
-                const next = e.target.value;
-                setCredentialDraft(next);
-                if (next.trim() === "") {
-                  setShowCredential(false);
-                }
-              }}
-            />
-            <button
-              type="button"
-              className={mmActionButtonClass({
-                variant: "tertiary",
-                disabled: !canOperate || panelBusy,
-              })}
-              disabled={!canOperate || panelBusy}
-              onClick={() => setShowCredential((v) => !v)}
-            >
-              {showCredential ? "Hide" : "Show"}
-            </button>
-          </div>
-          {hasInstance ? (
-            <p
-              id={`pruner-conn-key-hint-${provider}`}
-              className="text-xs text-[var(--mm-text3)]"
-            >
-              Leave blank to keep your saved{" "}
-              {provider === "plex" ? "token" : "key"}, or type a new one to
-              replace it.
-            </p>
-          ) : null}
-        </div>
-
-        {saveJustSucceeded ? (
-          <p
-            className="rounded-md border border-[rgba(212,175,55,0.45)] bg-[var(--mm-accent-soft)] px-3 py-2 text-sm font-medium text-[var(--mm-text1)]"
-            role="status"
-            data-testid={`pruner-connection-save-ok-${provider}`}
-          >
-            Saved.
-          </p>
-        ) : null}
-
-        {testJustSucceeded ? (
-          <p
-            className="rounded-md border border-[rgba(212,175,55,0.45)] bg-[var(--mm-accent-soft)] px-3 py-2 text-sm font-medium text-[var(--mm-text1)]"
-            role="status"
-            data-testid={`pruner-connection-test-ok-${provider}`}
-          >
-            Test started — results appear here when it finishes.
-          </p>
-        ) : null}
-      </div>
-
-      <div className="mm-card-action-footer mm-card-action-footer--row">
-        <button
-          type="button"
-          className={mmActionButtonClass({
-            variant: "primary",
-            disabled: !canOperate || !dirty || panelBusy,
-          })}
-          disabled={!canOperate || !dirty || panelBusy}
-          onClick={() => void saveConnection()}
-        >
-          {savePending ? "Saving…" : saveLabel}
-        </button>
-        <button
-          type="button"
-          className={mmActionButtonClass({
-            variant: "secondary",
-            disabled: !canOperate || panelBusy || !selectedInstance,
-          })}
-          disabled={!canOperate || panelBusy || !selectedInstance}
-          onClick={() => void runConnectionTest()}
-        >
-          {testPending ? "Testing…" : testLabel}
-        </button>
-      </div>
-
-      <div
-        className="mt-4 rounded-md border border-[var(--mm-border)] bg-[var(--mm-card-bg)] p-3.5 text-sm text-[var(--mm-text2)]"
-        data-testid={`pruner-connection-status-${provider}`}
-      >
-        <p className="text-sm font-medium text-[var(--mm-text1)]">
-          {connectionStatusMain}
-        </p>
-        <p className="mt-1 text-xs text-[var(--mm-text3)]">
-          Last completed check:{" "}
-          {formatPrunerDateTime(
-            selectedInstance?.last_connection_test_at ?? null,
-          )}
-        </p>
-        {selectedInstance?.last_connection_test_detail &&
-        selectedInstance.last_connection_test_ok !== true ? (
-          <p className="mt-1 text-xs text-[var(--mm-text3)]">
-            {selectedInstance.last_connection_test_detail}
-          </p>
-        ) : null}
-        {err ? (
-          <p className="mt-2 text-sm text-red-400" role="alert">
-            {err}
-          </p>
-        ) : null}
-        <p className="mt-2 text-xs text-[var(--mm-text3)]">
-          Each test adds a line to{" "}
-          <Link
-            to="/activity"
-            className="text-[var(--mm-accent)] underline-offset-2 hover:underline"
-          >
-            Activity
-          </Link>{" "}
-          for your records.
-        </p>
-      </div>
-    </section>
-  );
-}
-
-type ProviderWorkspaceSection = "connection" | "cleanup" | "schedule";
-
-function ProviderConfigurationWorkspace({
-  provider,
-  allInstances,
-  initialSection = "connection",
-}: {
-  provider: ProviderTab;
-  allInstances: PrunerServerInstance[];
-  initialSection?: ProviderWorkspaceSection;
-}) {
-  const providerName = providerLabel(provider);
-  const providerInstances = useMemo(
-    () => allInstances.filter((x) => x.provider === provider),
-    [allInstances, provider],
-  );
-  const [selectedInstanceId, setSelectedInstanceId] = useState<number | null>(
-    providerInstances[0]?.id ?? null,
-  );
-  const [providerSection, setProviderSection] =
-    useState<ProviderWorkspaceSection>(initialSection);
-  const selectedInstance =
-    providerInstances.find((x) => x.id === selectedInstanceId) ??
-    providerInstances[0];
-  const rulesCardRef = useRef<PrunerProviderRulesCardHandle>(null);
-
-  useEffect(() => {
-    setSelectedInstanceId((prev) => {
-      if (prev != null && providerInstances.some((x) => x.id === prev))
-        return prev;
-      return providerInstances[0]?.id ?? null;
-    });
-  }, [provider, providerInstances]);
-
-  useEffect(() => {
-    setProviderSection(initialSection);
-  }, [provider, initialSection]);
-
-  const disabledCtx = {
-    instanceId: 0,
-    instance: providerDisabledInstance(provider),
-  } as const;
-  const instanceSelection = {
-    selectedId: selectedInstanceId,
-    onSelectedIdChange: setSelectedInstanceId,
-  };
-
-  return (
-    <section
-      className="mm-bubble-stack"
-      data-testid={`pruner-provider-tab-${provider}`}
-    >
-      {providerInstances.length > 1 ? (
-        <label className="block max-w-md text-sm text-[var(--mm-text2)]">
-          <span className="mb-1 block text-xs text-[var(--mm-text3)]">
-            Server
-          </span>
-          <select
-            className="mm-input mt-1 w-full"
-            value={selectedInstance?.id ?? ""}
-            onChange={(e) => setSelectedInstanceId(Number(e.target.value))}
-          >
-            {providerInstances.map((inst) => (
-              <option key={inst.id} value={inst.id}>
-                {inst.display_name}
-              </option>
-            ))}
-          </select>
-        </label>
-      ) : null}
-
-      <nav
-        className="flex flex-wrap gap-2 border-b border-[var(--mm-border)] pb-3"
-        aria-label={`${providerName} configuration sections`}
-        data-testid={`pruner-provider-subnav-${provider}`}
-      >
-        {(
-          [
-            ["connection", "Connection"],
-            ["cleanup", "Cleanup"],
-            ["schedule", "Schedule"],
-          ] as const
-        ).map(([id, label]) => (
-          <button
-            key={id}
-            type="button"
-            className={mmSectionTabClass(providerSection === id)}
-            aria-current={providerSection === id ? "page" : undefined}
-            onClick={() => setProviderSection(id)}
-          >
-            {label}
-          </button>
-        ))}
-      </nav>
-
-      <div data-testid={`pruner-provider-sections-${provider}`}>
-        {providerSection === "connection" ? (
-          <PrunerConnectionCredentialPanel
-            provider={provider}
-            allInstances={allInstances}
-            instanceSelection={instanceSelection}
-          />
-        ) : null}
-
-        {providerSection === "cleanup" || providerSection === "schedule" ? (
-          <div
-            className={providerSection !== "cleanup" ? "hidden" : undefined}
-            data-testid={
-              providerSection === "cleanup"
-                ? "pruner-provider-cleanup-wrap"
-                : undefined
-            }
-            aria-hidden={providerSection !== "cleanup"}
-          >
-            <PrunerProviderRulesCard
-              ref={rulesCardRef}
-              provider={provider}
-              instanceId={selectedInstance?.id ?? 0}
-              instance={selectedInstance ?? disabledCtx.instance}
-            />
-          </div>
-        ) : null}
-
-        {providerSection === "schedule" ? (
-          <div
-            className="mm-dash-grid"
-            data-testid="pruner-provider-schedule-wrap"
-          >
-            <PrunerGlobalScheduleRow
-              provider={provider}
-              scope="tv"
-              instance={selectedInstance}
-              ensureScopeSaved={async () => {
-                await rulesCardRef.current?.ensureTvSaved();
-              }}
-            />
-            <PrunerGlobalScheduleRow
-              provider={provider}
-              scope="movies"
-              instance={selectedInstance}
-              ensureScopeSaved={async () => {
-                await rulesCardRef.current?.ensureMoviesSaved();
-              }}
-            />
-          </div>
-        ) : null}
-      </div>
-    </section>
-  );
-}
-
-function scanOutcomeReadable(o: string | null | undefined): string {
-  if (!o) return "—";
-  const u = o.toLowerCase();
-  if (u === "success") return "Finished OK";
-  if (u === "failed") return "Failed";
-  if (u === "unsupported") return "Not available";
-  return o;
-}
-
-function buildPrunerNeedsAttention(
-  instances: PrunerServerInstance[],
-): { text: string; tab: ProviderTab }[] {
-  const items: { text: string; tab: ProviderTab }[] = [];
-  if (instances.length === 0) {
-    items.push({
-      text: "No Emby, Jellyfin, or Plex servers connected — add one under Emby, Jellyfin, or Plex.",
-      tab: "emby",
-    });
-    return items;
-  }
-  for (const inst of instances) {
-    if (inst.last_connection_test_ok === false) {
-      const label = providerLabel(inst.provider as ProviderTab);
-      items.push({
-        text: `${label} (${inst.display_name}) connection test failed — check Connection tab.`,
-        tab: inst.provider as ProviderTab,
-      });
-    }
-  }
-  const allRulesZero = instances.every((inst) =>
-    inst.scopes.every((sc) => activeRuleCount(sc) === 0),
-  );
-  if (allRulesZero) {
-    items.push({
-      text: "No cleanup rules are enabled — turn on rules in the Cleanup tab.",
-      tab: "emby",
-    });
-  }
-  return items.slice(0, 8);
-}
-
-const PRUNER_ATTENTION_TAB_ORDER: ProviderTab[] = ["emby", "jellyfin", "plex"];
-
-function prunerAttentionOpenLabel(tab: ProviderTab): string {
-  return `Open ${providerLabel(tab)}`;
-}
-
-function PrunerLast30StatsTiles({
-  itemsRemoved,
-  previewRuns,
-  failedApplies,
-}: {
-  itemsRemoved: number;
-  previewRuns: number;
-  failedApplies: number;
-}) {
-  return (
-    <div>
-      <MmStatTileRow>
-        <MmStatTile label="Removed" value={itemsRemoved} />
-        <MmStatTile label="Preview" value={previewRuns} />
-        <MmStatTile label="Failed" value={failedApplies} />
-      </MmStatTileRow>
-      <MmStatCaption>Library cleanup activity · last 30 days</MmStatCaption>
-    </div>
-  );
-}
-
-const PRUNER_NEXT_STEPS_BODY =
-  "Use Emby, Jellyfin, or Plex to connect your media server and configure cleanup rules. Check Jobs for recent activity. Use Schedule for timed scans and optional hour limits.";
-
-function PrunerOverviewNextSteps({
-  onNavigate,
-}: {
-  onNavigate: (tab: TopTab) => void;
-}) {
-  return (
-    <MmOverviewSection
-      headingId="pruner-overview-next-steps-heading"
-      heading="Next steps"
-      data-testid="pruner-overview-next-steps"
-      data-overview-order="3"
-    >
-      <div className="mm-bubble-stack">
-        <p className="leading-relaxed">{PRUNER_NEXT_STEPS_BODY}</p>
-        <div className="flex flex-wrap gap-2.5 border-t border-[var(--mm-border)] pt-4">
-          <MmNextStepsButton label="Emby" onClick={() => onNavigate("emby")} />
-          <MmNextStepsButton
-            label="Jellyfin"
-            onClick={() => onNavigate("jellyfin")}
-          />
-          <MmNextStepsButton label="Plex" onClick={() => onNavigate("plex")} />
-          <MmNextStepsButton label="Jobs" onClick={() => onNavigate("jobs")} />
-          <MmNextStepsButton
-            label="Schedule"
-            onClick={() => onNavigate("schedule")}
-          />
-        </div>
-      </div>
-    </MmOverviewSection>
-  );
-}
-
-function PrunerOverviewNeedsAttention({
-  items,
-  onOpenProviderTab,
-}: {
-  items: { text: string; tab: ProviderTab }[];
-  onOpenProviderTab: (tab: ProviderTab) => void;
-}) {
-  const actionTabs = PRUNER_ATTENTION_TAB_ORDER.filter((t) =>
-    items.some((row) => row.tab === t),
-  );
-  return (
-    <MmOverviewSection
-      headingId="pruner-overview-needs-attention-heading"
-      heading="Needs attention"
-      data-testid="pruner-overview-needs-attention"
-      data-overview-order="2"
-    >
-      <MmNeedsAttentionList
-        items={items.map((row) => row.text)}
-        actions={
-          actionTabs.length > 0 ? (
-            <>
-              {actionTabs.map((tab) => (
-                <MmNextStepsButton
-                  key={tab}
-                  label={prunerAttentionOpenLabel(tab)}
-                  onClick={() => onOpenProviderTab(tab)}
-                />
-              ))}
-            </>
-          ) : undefined
-        }
-      />
-    </MmOverviewSection>
-  );
-}
-
-function TopLevelOverview({
-  instances,
-  onOpenProviderTab,
-  onNavigateTopTab,
-}: {
-  instances: PrunerServerInstance[];
-  onOpenProviderTab: (tab: ProviderTab) => void;
-  onNavigateTopTab: (tab: TopTab) => void;
-}) {
-  const jobsQ = usePrunerJobsInspectionQuery(50);
-  const statsQ = usePrunerOverviewStatsQuery();
-  const providers: ProviderTab[] = ["emby", "jellyfin", "plex"];
-  const providerCards = providers.map((provider) => {
-    const rows = instances.filter((x) => x.provider === provider);
-    const first = rows[0];
-    const scopeRows = first?.scopes ?? [];
-    const activeRules = scopeRows.reduce(
-      (acc, scope) => acc + activeRuleCount(scope),
-      0,
-    );
-    const previews = scopeRows.filter((scope) => scope.last_preview_at);
-    const latestPreview = previews.sort((a, b) =>
-      String(b.last_preview_at ?? "").localeCompare(
-        String(a.last_preview_at ?? ""),
-      ),
-    )[0];
-    const providerJobs = jobsQ.data?.jobs?.filter((j) => {
-      const sid = parseServerInstanceId(j);
-      return sid != null && rows.some((row) => row.id === sid);
-    });
-    const latestApplyLike =
-      providerJobs?.find((j) =>
-        String(j.job_kind).toLowerCase().includes("apply"),
-      ) ?? providerJobs?.[0];
-    return {
-      provider,
-      rows,
-      first,
-      activeRules,
-      latestPreview,
-      latestJob: latestApplyLike,
-    };
-  });
-
-  const attentionItems = buildPrunerNeedsAttention(instances);
-
-  const last30Body = statsQ.isPending ? (
-    <p className="text-[var(--mm-text3)]">Loading…</p>
-  ) : statsQ.isError ? (
-    <p className="text-red-400">{(statsQ.error as Error).message}</p>
-  ) : statsQ.data ? (
-    <PrunerLast30StatsTiles
-      itemsRemoved={statsQ.data.items_removed}
-      previewRuns={statsQ.data.preview_runs}
-      failedApplies={statsQ.data.failed_applies}
-    />
-  ) : (
-    <p className="text-[var(--mm-text3)]">—</p>
-  );
-
-  return (
-    <section className="mm-bubble-stack" data-testid="pruner-top-overview-tab">
-      <MmOverviewSection
-        headingId="pruner-overview-at-a-glance-heading"
-        heading="At a glance"
-        data-testid="pruner-overview-at-a-glance"
-        data-overview-order="1"
-      >
-        <MmAtGlanceGrid className="grid grid-cols-1 gap-4 sm:grid-cols-2 sm:gap-x-5 sm:gap-y-5 lg:grid-cols-12 lg:gap-x-5 lg:gap-y-6">
-          <MmAtGlanceCard
-            glanceOrder="1"
-            title="Last 30 days"
-            emphasis
-            body={last30Body}
-            gridClassName="lg:col-span-4"
-          />
-          {providerCards.map((card, i) => {
-            const order = String(i + 2) as "2" | "3" | "4";
-            const providerGridClass =
-              i === 2 ? "sm:col-span-2 lg:col-span-12" : "lg:col-span-4";
-            const body = card.first ? (
-              <div className="space-y-1.5">
-                <p>
-                  <span className="text-[var(--mm-text3)]">
-                    Connection test:
-                  </span>{" "}
-                  <span className="font-medium text-[var(--mm-text1)]">
-                    {card.first.last_connection_test_ok == null
-                      ? "Not run yet"
-                      : card.first.last_connection_test_ok
-                        ? "OK"
-                        : "Failed"}
-                  </span>
-                </p>
-                <p>
-                  <span className="text-[var(--mm-text3)]">
-                    Cleanup rules on:
-                  </span>{" "}
-                  <span className="font-medium text-[var(--mm-text1)]">
-                    {card.activeRules}
-                  </span>
-                </p>
-                <p>
-                  <span className="text-[var(--mm-text3)]">
-                    Last library scan:
-                  </span>{" "}
-                  <span className="font-medium text-[var(--mm-text1)]">
-                    {card.latestPreview
-                      ? `${card.latestPreview.media_scope === "tv" ? "TV" : "Movies"} · ${scanOutcomeReadable(card.latestPreview.last_preview_outcome)}`
-                      : "—"}
-                  </span>
-                </p>
-                <p>
-                  <span className="text-[var(--mm-text3)]">Recent task:</span>{" "}
-                  <span className="font-medium text-[var(--mm-text1)]">
-                    {card.latestJob
-                      ? `${prunerJobKindOperatorLabel(card.latestJob.job_kind)} (${card.latestJob.status})`
-                      : "—"}
-                  </span>
-                </p>
-              </div>
-            ) : (
-              <p className="text-[var(--mm-text2)]">
-                No server saved yet. Add the address and key on that provider’s
-                Connection tab.
-              </p>
-            );
-            return (
-              <MmAtGlanceCard
-                key={card.provider}
-                glanceOrder={order}
-                title={providerLabel(card.provider)}
-                body={body}
-                gridClassName={providerGridClass}
-                footer={
-                  card.provider === "emby" ? (
-                    <MmNextStepsButton
-                      label="Open Emby"
-                      onClick={() => onNavigateTopTab("emby")}
-                    />
-                  ) : card.provider === "jellyfin" ? (
-                    <MmNextStepsButton
-                      label="Open Jellyfin"
-                      onClick={() => onNavigateTopTab("jellyfin")}
-                    />
-                  ) : (
-                    <MmNextStepsButton
-                      label="Open Plex"
-                      onClick={() => onNavigateTopTab("plex")}
-                    />
-                  )
-                }
-              />
-            );
-          })}
-        </MmAtGlanceGrid>
-      </MmOverviewSection>
-
-      <PrunerOverviewNeedsAttention
-        items={attentionItems}
-        onOpenProviderTab={onOpenProviderTab}
-      />
-
-      <PrunerOverviewNextSteps onNavigate={onNavigateTopTab} />
-    </section>
-  );
-}
-
-function PrunerGlobalScheduleRow({
-  provider,
-  scope,
-  instance,
-  ensureScopeSaved,
-}: {
-  provider: ProviderTab;
-  scope: "tv" | "movies";
-  instance: PrunerServerInstance | undefined;
-  ensureScopeSaved: () => Promise<void>;
-}) {
-  const qc = useQueryClient();
-  const me = useMeQuery();
-  const canOperate = me.data?.role === "admin" || me.data?.role === "operator";
-  /** Draft UI uses defaults when no server exists yet; persist only once a real instance id is available. */
-  const displayInstance = instance ?? providerDisabledInstance(provider);
-  const scopeRow = displayInstance.scopes.find((s) => s.media_scope === scope);
-  const persistable = Boolean(instance && instance.id > 0);
-  const [schedHoursLimited, setSchedHoursLimited] = useState(false);
-  const [schedDays, setSchedDays] = useState("");
-  const [schedStart, setSchedStart] = useState("00:00");
-  const [schedEnd, setSchedEnd] = useState("23:59");
-  const [previewCap, setPreviewCap] = useState(500);
-  const [busy, setBusy] = useState(false);
-  const [msg, setMsg] = useState<string | null>(null);
-  const [err, setErr] = useState<string | null>(null);
-  const [dryRunEnabled, setDryRunEnabled] = useState(true);
-
-  useEffect(() => {
-    if (!scopeRow) {
-      setSchedHoursLimited(false);
-      setSchedDays("");
-      setSchedStart("00:00");
-      setSchedEnd("23:59");
-      setPreviewCap(500);
-      return;
-    }
-    setSchedHoursLimited(scopeRow.scheduled_preview_hours_limited ?? false);
-    setSchedDays(scopeRow.scheduled_preview_days ?? "");
-    setSchedStart(scopeRow.scheduled_preview_start ?? "00:00");
-    setSchedEnd(scopeRow.scheduled_preview_end ?? "23:59");
-    setPreviewCap(scopeRow.preview_max_items);
-  }, [
-    scopeRow?.scheduled_preview_hours_limited,
-    scopeRow?.scheduled_preview_days,
-    scopeRow?.scheduled_preview_start,
-    scopeRow?.scheduled_preview_end,
-    scopeRow?.preview_max_items,
-    instance?.id,
-  ]);
-
-  const scheduleFieldsDirty =
-    scopeRow != null &&
-    (schedHoursLimited !==
-      (scopeRow.scheduled_preview_hours_limited ?? false) ||
-      schedDays !== (scopeRow.scheduled_preview_days ?? "") ||
-      schedStart !== (scopeRow.scheduled_preview_start ?? "00:00") ||
-      schedEnd !== (scopeRow.scheduled_preview_end ?? "23:59") ||
-      previewCap !== scopeRow.preview_max_items);
-
-  async function saveRow() {
-    if (!scopeRow) return;
-    if (!persistable) {
-      setMsg(null);
-      setErr(
-        "Save a server on the Connection tab first, then you can save this schedule to that server.",
-      );
-      return;
-    }
-    setMsg(null);
-    setErr(null);
-    setBusy(true);
-    try {
-      const csrf_token = await fetchCsrfToken();
-      const cap = Math.max(1, Math.min(5000, Number(previewCap) || 500));
-      await patchPrunerScope(instance!.id, scope, {
-        scheduled_preview_enabled: scopeRow.scheduled_preview_enabled,
-        scheduled_preview_interval_seconds:
-          scopeRow.scheduled_preview_interval_seconds,
-        scheduled_preview_hours_limited: schedHoursLimited,
-        scheduled_preview_days: schedDays,
-        scheduled_preview_start: schedStart,
-        scheduled_preview_end: schedEnd,
-        preview_max_items: cap,
-        csrf_token,
-      });
-      await qc.invalidateQueries({ queryKey: ["pruner", "instances"] });
-      setMsg("Saved.");
-    } catch (e) {
-      setErr((e as Error).message);
-    } finally {
-      setBusy(false);
-    }
-  }
-
-  const controlsDisabled = !canOperate || busy;
-  const saveDisabled = busy || !canOperate || !scopeRow || !scheduleFieldsDirty;
-  const testId = `pruner-schedule-row-${provider}-${scope}`;
-  const idPrefix = `pruner-sched-${provider}-${scope}`;
-
-  const saveScheduleLabel =
-    scope === "tv" ? "Save TV schedule" : "Save Movies schedule";
-
-  const laneTitle =
-    scope === "tv"
-      ? "TV automatic scan window"
-      : "Movies automatic scan window";
-  const laneIntro =
-    scope === "tv"
-      ? "Limit which days and times Pruner may run scheduled TV cleanup previews for this provider."
-      : "Limit which days and times Pruner may run scheduled Movies cleanup previews for this provider.";
-
-  return (
-    <section
-      className="mm-card mm-dash-card flex h-full min-h-0 min-w-0 flex-col p-6"
-      data-testid={testId}
-    >
-      <div className="mm-card-action-body flex-1 min-h-0">
-        <div>
-          <h3 className="text-base font-semibold text-[var(--mm-text1)]">
-            {laneTitle}
-          </h3>
-          <p className="mt-1 text-sm text-[var(--mm-text2)]">{laneIntro}</p>
-        </div>
-        <div className="space-y-3">
-          <div>
-            <span className="text-sm font-medium text-[var(--mm-text1)]">
-              Schedule window
-            </span>
-            <p className="mt-1 text-xs text-[var(--mm-text3)]">
-              {MM_SCHEDULE_TIME_WINDOW_HELPER}
-            </p>
-          </div>
-          <div className="space-y-4">
-            <MmOnOffSwitch
-              id={`${idPrefix}-hours-limited`}
-              label="Limit to these hours"
-              enabled={schedHoursLimited}
-              disabled={controlsDisabled}
-              onChange={setSchedHoursLimited}
-            />
-            <div className="space-y-2">
-              <span className="text-sm font-medium text-[var(--mm-text1)]">
-                Days
-              </span>
-              <MmScheduleDayChips
-                scheduleDaysCsv={schedDays}
-                disabled={controlsDisabled}
-                onChangeCsv={setSchedDays}
-              />
-            </div>
-            <MmScheduleTimeFields
-              idPrefix={idPrefix}
-              start={schedStart}
-              end={schedEnd}
-              disabled={controlsDisabled}
-              onStart={setSchedStart}
-              onEnd={setSchedEnd}
-            />
-          </div>
-        </div>
-        <div>
-          <span className="text-sm font-medium text-[var(--mm-text1)]">
-            Items to scan per run
-          </span>
-          <p className="mt-1 text-xs text-[var(--mm-text3)]">
-            How many items to check each time the scan runs. Higher numbers take
-            longer. Maximum 5,000.
-          </p>
-          <input
-            type="number"
-            min={1}
-            max={5000}
-            className="mm-input mt-2 w-full"
-            value={previewCap}
-            disabled={controlsDisabled}
-            onChange={(e) =>
-              setPreviewCap(
-                Math.max(1, Math.min(5000, Number(e.target.value) || 500)),
-              )
-            }
-            aria-label="Items to scan per run"
-          />
-        </div>
-        <p className="text-xs text-[var(--mm-text3)]">
-          Last automatic scan:{" "}
-          <span className="font-medium text-[var(--mm-text1)]">
-            {scopeRow?.last_scheduled_preview_enqueued_at
-              ? formatPrunerDateTime(
-                  scopeRow.last_scheduled_preview_enqueued_at,
-                )
-              : "Never run"}
-          </span>
-        </p>
-        {!canOperate && msg ? (
-          <p className="text-xs text-green-600">{msg}</p>
-        ) : null}
-        {!canOperate && err ? (
-          <p className="text-xs text-red-500" role="alert">
-            {err}
-          </p>
-        ) : null}
-      </div>
-      {canOperate ? (
-        <div className="mm-card-action-footer">
-          <button
-            type="button"
-            className={`${mmActionButtonClass({ variant: "primary", disabled: saveDisabled })} w-full`}
-            disabled={saveDisabled}
-            onClick={() => void saveRow()}
-          >
-            {busy ? "Saving…" : saveScheduleLabel}
-          </button>
-          {msg ? <p className="text-xs text-green-600">{msg}</p> : null}
-          {err ? (
-            <p className="text-xs text-red-500" role="alert">
-              {err}
-            </p>
-          ) : null}
-        </div>
-      ) : null}
-
-      <div className="mt-4 border-t border-[var(--mm-border)] pt-6">
-        <h4 className="text-base font-semibold text-[var(--mm-text1)]">
-          Run now
-        </h4>
-        <p className="mt-1 text-xs text-[var(--mm-text3)]">
-          Scan your saved cleanup criteria immediately without waiting for the
-          schedule. This creates review snapshots; deletion still requires a
-          saved snapshot confirmation.
-        </p>
-        <div className="mt-4">
-          <PrunerDryRunControls
-            instanceId={instance?.id ?? 0}
-            mediaScope={scope}
-            testIdPrefix={`pruner-schedule-${provider}`}
-            ensureSaved={ensureScopeSaved}
-            dryRunEnabled={dryRunEnabled}
-            onDryRunEnabledChange={setDryRunEnabled}
-            runDisabled={!instance || instance.id <= 0}
-            controlsDisabled={!canOperate || busy}
-          />
-        </div>
-      </div>
-    </section>
-  );
-}
-
-function TopLevelJobs({ instances }: { instances: PrunerServerInstance[] }) {
-  const PAGE_SIZE_OPTIONS = [20, 50, 100] as const;
-  const jobsQ = usePrunerJobsInspectionQuery(100);
-  const byId = useMemo(
-    () => new Map(instances.map((x) => [x.id, x])),
-    [instances],
-  );
-  const [statusFilter, setStatusFilter] = useState("recent");
-  const [page, setPage] = useState(1);
-  const [pageSize, setPageSize] = useState<number>(PAGE_SIZE_OPTIONS[0]);
-  const filterLabelId = useId();
-  const jobs =
-    statusFilter === "recent"
-      ? (jobsQ.data?.jobs ?? [])
-      : (jobsQ.data?.jobs ?? []).filter((j) => j.status === statusFilter);
-  const totalPages = Math.max(1, Math.ceil(jobs.length / pageSize));
-  const pagedRows = jobs.slice((page - 1) * pageSize, page * pageSize);
-
-  useEffect(() => {
-    setPage(1);
-  }, [statusFilter, pageSize]);
-
-  useEffect(() => {
-    if (page > totalPages) {
-      setPage(totalPages);
-    }
-  }, [page, totalPages]);
-  return (
-    <section
-      className="rounded-lg border border-[var(--mm-border)] bg-[var(--mm-card-bg)] shadow-sm"
-      data-testid="pruner-top-jobs-tab"
-    >
-      <header className="border-b border-[var(--mm-border)] bg-black/10 px-4 py-3.5 sm:px-5 sm:py-4">
-        <h2 className="text-lg font-semibold tracking-tight text-[var(--mm-text)]">
-          Jobs
-        </h2>
-        <p className="mt-1 text-sm text-[var(--mm-text2)]">
-          Pending, running, and recent Pruner work.
-        </p>
-      </header>
-      <div className="space-y-4 px-4 py-4 sm:px-5 sm:py-5">
-        <div className="flex flex-col gap-3 rounded-md border border-[var(--mm-border)] bg-[var(--mm-card-bg)] px-3.5 py-3.5 sm:flex-row sm:items-end sm:justify-between sm:px-5 sm:py-4">
-          <label className="block min-w-0 flex-1">
-            <span
-              id={filterLabelId}
-              className="text-xs font-semibold uppercase tracking-wide text-[var(--mm-text3)]"
-            >
-              Show jobs
-            </span>
-            <MmListboxPicker
-              className="mt-2 max-w-xl"
-              ariaLabelledBy={filterLabelId}
-              placeholder="Recent (all statuses, newest first)"
-              options={PRUNER_JOB_FILTER_OPTIONS.map((o) => ({
-                value: o.value,
-                label: o.label,
-              }))}
-              value={statusFilter}
-              onChange={(v) => setStatusFilter(v)}
-            />
-          </label>
-        </div>
-        {jobsQ.isLoading ? (
-          <p className="text-sm text-[var(--mm-text2)]">Loading jobs…</p>
-        ) : null}
-        {jobsQ.isError ? (
-          <p className="text-sm text-red-600">
-            {(jobsQ.error as Error).message}
-          </p>
-        ) : null}
-        {jobs.length ? (
-          <>
-            <div className="overflow-x-auto rounded-md border border-[var(--mm-border)]">
-              <table className="w-full min-w-[32rem] border-collapse text-left text-sm">
-                <thead className="bg-black/20 text-[var(--mm-text2)]">
-                  <tr>
-                    <th className="sticky left-0 top-0 z-30 bg-black/20 px-3 py-2 font-medium">
-                      Id
-                    </th>
-                    <th className="sticky top-0 z-20 bg-black/20 px-3 py-2 font-medium">
-                      What ran
-                    </th>
-                    <th className="sticky top-0 z-20 bg-black/20 px-3 py-2 font-medium">
-                      Status
-                    </th>
-                    <th className="sticky top-0 z-20 bg-black/20 px-3 py-2 font-medium">
-                      Server
-                    </th>
-                    <th className="sticky top-0 z-20 bg-black/20 px-3 py-2 font-medium">
-                      Updated
-                    </th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {pagedRows.map((job) => {
-                    const sid = parseServerInstanceId(job);
-                    const inst = sid ? byId.get(sid) : undefined;
-                    return (
-                      <tr
-                        key={job.id}
-                        className="border-t border-[var(--mm-border)]"
-                      >
-                        <td className="sticky left-0 z-[1] bg-[var(--mm-card-bg)] px-3 py-2 align-top font-mono text-xs text-[var(--mm-text1)]">
-                          #{job.id}
-                        </td>
-                        <td className="max-w-[14rem] break-words px-3 py-2 align-top text-xs">
-                          {prunerJobKindOperatorLabel(job.job_kind)}
-                        </td>
-                        <td className="whitespace-nowrap px-3 py-2 align-top">
-                          <span
-                            className={
-                              job.status === "completed"
-                                ? "text-emerald-500"
-                                : job.status === "failed"
-                                  ? "text-red-400"
-                                  : job.status === "running"
-                                    ? "text-amber-400"
-                                    : "text-[var(--mm-text2)]"
-                            }
-                          >
-                            {job.status}
-                          </span>
-                        </td>
-                        <td className="max-w-[14rem] break-words px-3 py-2 align-top text-xs">
-                          {inst
-                            ? inst.display_name
-                            : sid
-                              ? `Server #${sid}`
-                              : "—"}
-                        </td>
-                        <td className="whitespace-nowrap px-3 py-2 align-top text-xs text-[var(--mm-text2)]">
-                          {job.updated_at}
-                        </td>
-                      </tr>
-                    );
-                  })}
-                </tbody>
-              </table>
-            </div>
-            <MmJobsPagination
-              page={page}
-              totalPages={totalPages}
-              onPageChange={setPage}
-              pageSize={pageSize}
-              onPageSizeChange={setPageSize}
-              pageSizeOptions={[...PAGE_SIZE_OPTIONS]}
-            />
-          </>
-        ) : jobsQ.data ? (
-          <div className="space-y-1 rounded border border-[var(--mm-border)] bg-black/10 px-5 py-10 text-center">
-            <p className="text-sm font-medium text-[var(--mm-text)]">
-              No jobs match this view
-            </p>
-            <p className="text-xs text-[var(--mm-text2)]">
-              {statusFilter !== "recent"
-                ? `Nothing with status "${statusFilter}" yet. Try Recent (all statuses) for the latest rows.`
-                : "No recent Pruner jobs yet. Run cleanup or connection tasks to see activity here."}
-            </p>
-          </div>
-        ) : null}
-        <p className="text-xs text-[var(--mm-text2)]">
-          Full detail on what was deleted, skipped, or failed is in the{" "}
-          <Link
-            to="/activity"
-            className="font-semibold text-[var(--mm-accent)] underline-offset-2 hover:underline"
-          >
-            Activity log
-          </Link>
-          .
-        </p>
-      </div>
-    </section>
-  );
-}
+import { mmSectionTabClass } from "../../lib/ui/mm-control-roles";
+import { usePrunerInstancesQuery } from "../../lib/pruner/queries";
+import { PRUNER_TAB_BLURBS } from "./pruner-page-constants";
+import type { TopTab } from "./pruner-page-types";
+import { ProviderConfigurationWorkspace } from "./pruner-provider-configuration-workspace";
+import { TopLevelJobs } from "./pruner-top-level-jobs";
+import { TopLevelOverview } from "./pruner-top-level-overview";
 
 export function PrunerInstancesListPage() {
   const q = usePrunerInstancesQuery();
   const [topTab, setTopTab] = useState<TopTab>("overview");
   const instances = q.data ?? [];
+
   return (
     <div className="mm-page w-full min-w-0" data-testid="pruner-scope-page">
       <header className="mm-page__intro !mb-0">
@@ -1441,9 +48,7 @@ export function PrunerInstancesListPage() {
             key={id}
             type="button"
             role="tab"
-            aria-selected={
-              topTab === id || (topTab === "schedule" && id === "emby")
-            }
+            aria-selected={topTab === id || (topTab === "schedule" && id === "emby")}
             className={mmSectionTabClass(
               topTab === id || (topTab === "schedule" && id === "emby"),
             )}
@@ -1470,18 +75,13 @@ export function PrunerInstancesListPage() {
           )[topTab]
         }
       >
-        <div
-          className={mmModuleTabBlurbBandClass}
-          data-testid="pruner-tab-blurb"
-        >
-          <p className={mmModuleTabBlurbTextClass}>
-            {PRUNER_TAB_BLURBS[topTab]}
-          </p>
+        <div className={mmModuleTabBlurbBandClass} data-testid="pruner-tab-blurb">
+          <p className={mmModuleTabBlurbTextClass}>{PRUNER_TAB_BLURBS[topTab]}</p>
         </div>
         <div className="min-w-0">
           {q.isLoading ? (
             <p className="text-sm text-[var(--mm-text2)]">
-              Loading provider instances…
+              Loading provider instances...
             </p>
           ) : null}
           {q.isError ? (
@@ -1503,10 +103,7 @@ export function PrunerInstancesListPage() {
                 initialSection="schedule"
               />
             ) : (
-              <ProviderConfigurationWorkspace
-                provider={topTab}
-                allInstances={instances}
-              />
+              <ProviderConfigurationWorkspace provider={topTab} allInstances={instances} />
             )
           ) : null}
         </div>
