+++ apps/web/src/pages/pruner/pruner-page-types.ts
@@ -1,0 +1,11 @@
+export type TopTab =
+  | "overview"
+  | "emby"
+  | "jellyfin"
+  | "plex"
+  | "jobs"
+  | "schedule";
+
+export type ProviderTab = "emby" | "jellyfin" | "plex";
+
+export type ProviderWorkspaceSection = "connection" | "cleanup" | "schedule";
